Return Error::ResponseError for API errors
Due to a bug in the error handling code, we previously returned
Error::Transport for API errors.  With this patch, we correctly return
Error::ResponseError.  This also means that the error message from the
NetHSM can be accessed using the content field of the ResponseContent.
Unfortunately, it is not yet deserialized as it is not described in the
API spec.

Fixes: https://github.com/Nitrokey/nethsm-sdk-rs/issues/21